api/requirements.txt
fastapi==0.110.*
uvicorn==0.27.*
sqlmodel==0.0.16
pydantic==2.11.*
pydantic-settings==2.5.*
python-jose==3.3.*
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
aiosqlite==0.19.*
alembic==1.13.*
PyYAML==6.0.*
python-dotenv==1.0.*
greenlet==3.0.*
